backend/requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.12
cryptography==41.0.2
defusedxml==0.7.1
django-filter==23.1
Django==3.2.3
djangorestframework==3.12.4
djoser==2.1.0
gunicorn==20.1.0
Pillow==9.0.0
psycopg2-binary==2.9.3
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytest==6.2.4
python-dotenv==0.19.0
PyYAML==6.0
requests==2.26.0
sentry-sdk==1.16.0
typing_extensions==4.7.1
urllib3==1.26.16
webcolors==1.11.1